La verification du token et la fonctionnalite de deconnexion reussi

+++ controller/user.go
@@ -48,6 +48,10 @@ func LoginHandler(w http.ResponseWriter, r *http.Request) {
 		return
 	}
 
+	user.Token = token
+    // Mettre à jour le modèle User dans la base de données avec le nouveau token
+    model.DB.Save(&user)
+
 	utils.RespondWithJSON(w, http.StatusOK, "Login successful", map[string]string{"token": token})
 }
 
@@ -72,33 +76,38 @@ func GetUserById(w http.ResponseWriter, r *http.Request) {
 }
 
 
-func UpdateUser(w http.ResponseWriter, r *http.Request){
+func UpdateUser(w http.ResponseWriter, r *http.Request) {
 	userUpdate := &model.User{}
 	utils.ParseBody(r, &userUpdate)
 	userId := chi.URLParam(r, "userId")
 
-	userDetail, db := model.GetUserById(userId)
-
-	if userDetail == nil {
-		http.Error(w, "User not found", http.StatusNotFound)
+	// Vérifier le token avant de continuer
+	if !provider.VerificationToken(r, userId) {
+		http.Error(w, "Unauthorized: Invalid token", http.StatusUnauthorized)
 		return
 	}
 
+	userDetail, db := model.GetUserById(userId)
+	if userDetail == nil {
+			http.Error(w, "User not found", http.StatusNotFound)
+			return
+	}
+
 	if userUpdate.Name != "" {
-		userDetail.Name = userUpdate.Name
+			userDetail.Name = userUpdate.Name
 	}
 
 	if userUpdate.UserName != "" {
-		userDetail.UserName = userUpdate.UserName
+			userDetail.UserName = userUpdate.UserName
 	}
 
 	if userUpdate.Email != "" {
-		userDetail.Email = userUpdate.Email
+			userDetail.Email = userUpdate.Email
 	}
 
-	if userDetail.Password != ""{
-		hashedPassword, _ := utils.HashPassword(userUpdate.Password)
-		userDetail.Password = hashedPassword
+	if userUpdate.Password != "" {
+			hashedPassword, _ := utils.HashPassword(userUpdate.Password)
+			userDetail.Password = hashedPassword
 	}
 
 	db.Save(&userDetail)
@@ -109,12 +118,40 @@ func UpdateUser(w http.ResponseWriter, r *http.Request){
 	w.Write(res)
 }
 
-func DeleteUser(w http.ResponseWriter, r *http.Request){
-	userId := chi.URLParam(r, "userId")
-	u := model.DeleteUser(userId)
 
+func DeleteUser(w http.ResponseWriter, r *http.Request) {
+	userId := chi.URLParam(r, "userId")
+
+	if !provider.VerificationToken(r, userId) {
+		http.Error(w, "Unauthorized: Invalid token", http.StatusUnauthorized)
+		return
+	}
+
+	u := model.DeleteUser(userId)
 	res, _ := json.Marshal(u)
 	w.Header().Set("content-type", "application/json")
 	w.WriteHeader(http.StatusOK)
 	w.Write(res)
+}
+
+
+
+func LogoutUser(w http.ResponseWriter, r *http.Request) {
+	userId := chi.URLParam(r, "userId")
+	u, _ := model.GetUserById(userId)
+
+	if u == nil {
+			http.Error(w, "User not found", http.StatusNotFound)
+			return
+	}
+
+	if !provider.VerificationToken(r, userId) {
+		http.Error(w, "Unauthorized: Invalid token", http.StatusUnauthorized)
+		return
+	}
+
+	// Révoquer le token en effaçant celui enregistré dans la base de données
+	u.Logout()
+
+	utils.RespondWithJSON(w, http.StatusOK, "Logout successful", nil)
 }
